refactor : entity 패키지 생성
