+++ src/main/java/homeat/backend/domain/post/entity/FoodTalkComment.java
@@ -1,11 +1,9 @@
-package homeat.backend.domain.post;
+package homeat.backend.domain.post.entity;
 
 import homeat.backend.global.common.domain.BaseEntity;
 import homeat.backend.user.domain.Member;
 import javax.persistence.Column;
 import javax.persistence.Entity;
-import javax.persistence.EnumType;
-import javax.persistence.Enumerated;
 import javax.persistence.FetchType;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
@@ -23,37 +21,20 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @Builder
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class FoodTalk extends BaseEntity {
+public class FoodTalkComment extends BaseEntity {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "foodtalk_id")
+    @Column(name = "foodtalk_comment_id")
     private Long id;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "foodtalk_id")
+    private FoodTalk foodTalk;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "member_id")
     private Member member;
 
-    private String name;
-    private String ingredient;
-    private String memo;
-
-    @Enumerated(EnumType.STRING)
-    private Tag tag;
-
-    private Integer love;
-    private Integer view;
-    private Integer commentNumber;
-
-    @Enumerated(EnumType.STRING)
-    private Save save;
-
-    public void update(String name, String ingredient, String memo, Tag tag) {
-        this.name = name;
-        this.ingredient = ingredient;
-        this.memo = memo;
-        this.tag = tag;
-    }
-
-
+    private String content;
 }
