+++ src/main/java/homeat/backend/domain/post/entity/InfoHashTag.java
@@ -1,7 +1,5 @@
-package homeat.backend.domain.post;
+package homeat.backend.domain.post.entity;
 
-import homeat.backend.global.common.domain.BaseEntity;
-import homeat.backend.user.domain.Member;
 import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.persistence.FetchType;
@@ -21,20 +19,16 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @Builder
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class InfoTalkComment extends BaseEntity {
+public class InfoHashTag {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "infotalk_comment_id")
+    @Column(name = "hashtag_id")
     private Long id;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "infotalk_id")
     private InfoTalk infoTalk;
 
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "member_id")
-    private Member member;
-
-    private String content;
+    private String tag;
 }
