+++ src/main/java/homeat/backend/domain/post/entity/InfoTalk.java
@@ -1,9 +1,11 @@
-package homeat.backend.domain.post;
+package homeat.backend.domain.post.entity;
 
 import homeat.backend.global.common.domain.BaseEntity;
 import homeat.backend.user.domain.Member;
 import javax.persistence.Column;
 import javax.persistence.Entity;
+import javax.persistence.EnumType;
+import javax.persistence.Enumerated;
 import javax.persistence.FetchType;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
@@ -21,20 +23,25 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @Builder
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class FoodTalkComment extends BaseEntity {
+public class InfoTalk extends BaseEntity {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "foodtalk_comment_id")
+    @Column(name = "infotalk_id")
     private Long id;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "foodtalk_id")
-    private FoodTalk foodTalk;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "member_id")
     private Member member;
 
+    private String title;
     private String content;
+
+
+    private Integer love;
+    private Integer view;
+    private Integer commentNumber;
+
+    @Enumerated(EnumType.STRING)
+    private Save save;
 }
