+++ src/main/java/homeat/backend/domain/post/entity/Save.java
@@ -1,4 +1,4 @@
-package homeat.backend.domain.post;
+package homeat.backend.domain.post.entity;
 
 public enum Save {
     임시저장, 저장
