+++ src/main/java/homeat/backend/domain/post/entity/Tag.java
@@ -1,4 +1,4 @@
-package homeat.backend.domain.post;
+package homeat.backend.domain.post.entity;
 
 public enum Tag {
     아침, 점심, 저녁
