+++ src/main/java/homeat/backend/domain/post/repository/FoodPictureRepository.java
@@ -1,6 +1,6 @@
 package homeat.backend.domain.post.repository;
 
-import homeat.backend.domain.post.FoodPicture;
+import homeat.backend.domain.post.entity.FoodPicture;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
