+++ src/main/java/homeat/backend/domain/post/repository/FoodTalkCommentRepository.java
@@ -1,6 +1,6 @@
 package homeat.backend.domain.post.repository;
 
-import homeat.backend.domain.post.FoodTalkComment;
+import homeat.backend.domain.post.entity.FoodTalkComment;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
