+++ src/main/java/homeat/backend/domain/post/repository/FoodTalkRepository.java
@@ -1,6 +1,6 @@
 package homeat.backend.domain.post.repository;
 
-import homeat.backend.domain.post.FoodTalk;
+import homeat.backend.domain.post.entity.FoodTalk;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
