+++ src/main/java/homeat/backend/domain/post/repository/InfoHashTagRepository.java
@@ -1,6 +1,6 @@
 package homeat.backend.domain.post.repository;
 
-import homeat.backend.domain.post.InfoHashTag;
+import homeat.backend.domain.post.entity.InfoHashTag;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
