+++ src/main/java/homeat/backend/domain/post/repository/InfoPictureRepository.java
@@ -1,6 +1,6 @@
 package homeat.backend.domain.post.repository;
 
-import homeat.backend.domain.post.InfoPicture;
+import homeat.backend.domain.post.entity.InfoPicture;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
