+++ src/main/java/homeat/backend/domain/post/repository/InfoTalkCommentRepository.java
@@ -1,6 +1,6 @@
 package homeat.backend.domain.post.repository;
 
-import homeat.backend.domain.post.InfoTalkComment;
+import homeat.backend.domain.post.entity.InfoTalkComment;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
