+++ src/main/java/homeat/backend/domain/post/repository/InfoTalkRepository.java
@@ -1,6 +1,6 @@
 package homeat.backend.domain.post.repository;
 
-import homeat.backend.domain.post.InfoTalk;
+import homeat.backend.domain.post.entity.InfoTalk;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
